pyproject.toml
[tool.poetry]
name = "peewee-test"
version = "0.1.0"
description = ""
authors = ["Vitalii Bielimov <vbit2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
peewee = "^3.17.1"


[tool.poetry.group.dev.dependencies]
aiohttp = "3.9.3"
aiosignal = "1.3.1"
alembic = "1.13.1"
async-timeout = "4.0.3"
attrs = "23.2.0"
faker = "23.2.0"
frozenlist = "1.4.1"
idna = "3.6"
mako = "1.3.2"
markupsafe = "2.1.5"
multidict = "6.0.5"
peewee = "3.17.1"
psycopg2-binary = "2.9.9"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
six = "1.16.0"
sqlalchemy = "2.0.27"
typing-extensions = "4.9.0"
yarl = "1.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
